Task 1 - KNN model done, LR model done, SVR in progress.
The library "caret" is used to create each training model.
Since all methods are integrated with this it has made it simple to just make
a few adjustments to the parameters of the traning model to better fit the
predictions. (May cause data bias, but it is not relevant for the given assignment)